Fix markup on "Building on Linux"
The table of optional dependencies didn't render on github.
I missed this when editing it because my local markdown renderer
looked fine. Hopefully fixed now.

In the table, I merged the two columns for dependencies under
debian and fedora into a single column, annotating the single entry
which applies to fedora. This might become unwieldy if the table gets
bulked out with complete dependency lists for many distros, but we'll
figure the best solution to that if and when it happens.

While I was in there, I also moved the final '8. Variations' section,
which only described setting a config variable to define the install
directory, up into the section about setting config variables.